Improve wpt testing on OffscreenCanvas convertToBlob
Same as canvas.toBlob(), the OffscreenCanvas.convertToBlob() are running on
idle tasks by default. When idle tasks do not begin for a long time, they force
encoding on the current thread. As idle tasks are not enabled on layout tests
unless "--enable-threaded-compositing" flag is turned on, we therefore do both
virtual and non-virtual tests to ensure complete test coverage of both
execution paths--idle-task mode and force-encoding mode. The non-virtual tests
need to be marked as [Slow] as the CanvasAsyncBlobCreator is waiting for the
idle tasks for a threshold time before switching to force-encoding mode.

Bug: 817091
Change-Id: I2c7da9e27ee2d24e1a5e6ee1c36597b46f052aca